Split apply logic by fragment type
Remove the Applier type and replace it with TextApplier and
BinaryApplier, both of which operate on fragements instead of on full
files. Move the logic that previously existed in Applier.ApplyFile to
the top-level Apply function.

Also restructure arguments and methods to make it clear that appliers
are one-time-use objects. The destination is now set when creating an
applier and the Reset() method was replaced by Close().
